Move backup embed to /embeds/ folder as well
